Move content in models to types
