+++ types/spoonacular/ErrorResponse.ts
