+++ types/spoonacular/SearchResponse.ts
